requirements.txt
ruamel.yaml
easydict
click
rich
pyfiglet
pytest